Move form processor ledger models to module
